fix : CommentItem.tsx -> data={data} 추가 refactoring : - view (Layout, Nav, Search) - Layout : css file naminge change (viewLayout) - Nav : css file naming change (layoutNav)
Relate to : #124
