+++ src/apis/community/getUser.ts
@@ -6,7 +6,9 @@ export const getUser = async () => {
 		data: { user },
 		error,
 	} = await supabase.auth.getUser();
-	if (error) throw error;
+	if (error) {
+		throw error;
+	}
 	const user_id = user?.id;
 	const user_email = user?.email;
 	const user_name = user?.user_metadata.name;
