+++ src/app/community/(view)/layout.tsx
@@ -1,9 +1,9 @@
-import Nav from '@/components/community/view/Nav';
+import Nav from '@/components/community/view/nav/Nav';
 import Search from '@/components/community/view/Search';
 import { BasicLayoutType } from '@/types/layoutType';
 import type { Metadata } from 'next';
 import Link from 'next/link';
-import styles from '@/styles/community/layout.module.css';
+import styles from '@/styles/community/ViewLayout.module.css';
 import Header from '@/components/community/common/Header';
 import { AllDataType } from '@/types/community/view/data';
 import { getDate } from '@/utils/getDate';
@@ -12,27 +12,29 @@ export const metadata: Metadata = {
 	description: '',
 };
 
-export default async function CommunityLayout({ children }: BasicLayoutType) {
+export default async function CommunityViewLayout({
+	children,
+}: BasicLayoutType) {
 	const reponse = await fetch('http://localhost:8080/popular/community', {
 		cache: 'no-store',
 	});
 	const popularData = await reponse.json();
 
 	return (
-		<>
+		<main>
 			<Header />
 			<div className={styles.container}>
 				{/* 왼쪽 광고 영역 */}
-				<div></div>
+				<aside></aside>
 				<div className={styles.mainWrap}>
-					<div className={styles.contentWrap}>
+					<article className={styles.contentWrap}>
 						<Nav />
 						<Search />
 						<hr className={styles.line} />
 						{children}
-					</div>
+					</article>
 
-					<div className={styles.weeklyPopularWrap}>
+					<aside className={styles.weeklyPopularWrap}>
 						<div>
 							<h3>주간 인기글</h3>
 
@@ -54,11 +56,11 @@ export default async function CommunityLayout({ children }: BasicLayoutType) {
 								);
 							})}
 						</div>
-					</div>
+					</aside>
 				</div>
 				{/* 오른쪽 광고 영역 */}
-				<div></div>
+				<aside></aside>
 			</div>
-		</>
+		</main>
 	);
 }
