--- src/styles/community/nav.module.css
@@ -1,42 +1,0 @@
-.container {
-	width: 100%;
-	display: flex;
-	justify-content: space-between;
-}
-/* Link button */
-.linkWrap {
-	display: flex;
-	align-items: center;
-}
-
-.linkItem {
-	margin-right: 10px;
-	width: 100px;
-	height: 45px;
-	text-align: center;
-	line-height: 45px;
-	cursor: pointer;
-}
-.linkItem:hover {
-	background-color: #6594e7;
-	color: #ffffff;
-}
-.active {
-	margin-right: 10px;
-	width: 100px;
-	height: 45px;
-	text-align: center;
-	line-height: 45px;
-	background-color: #6594e7;
-	color: #ffffff;
-	cursor: pointer;
-}
-/* write button */
-.writeBtn {
-	width: 100px;
-	height: 45px;
-	text-align: center;
-	line-height: 45px;
-	background-color: #6594e7;
-	color: #ffffff;
-}
